python-postgres-project/requirements.txt
docker
psycopg2-binary
docker
docker-compose